requirements.txt
streamlit==1.32.0
google-genai==1.47.0
python-dotenv==1.0.1